Struct rusoto_marketplace_entitlement::MarketplaceEntitlementClient

pub struct MarketplaceEntitlementClient { /* fields omitted */ }

A client for the AWS Marketplace Entitlement Service API.

Methods

impl MarketplaceEntitlementClient

pub fn new(region: Region) -> MarketplaceEntitlementClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> MarketplaceEntitlementClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,