Crate rusoto_marketplace_entitlement

AWS Marketplace Entitlement Service
This reference provides descriptions of the AWS Marketplace Entitlement Service API.

AWS Marketplace Entitlement Service is used to determine the entitlement of a customer to a given product. An entitlement represents capacity in a product owned by the customer. For example, a customer might own some number of users or seats in an SaaS application or some amount of data capacity in a multi-tenant database.

Getting Entitlement Records

GetEntitlements- Gets the entitlements for a Marketplace product.

If you're using the service, you're probably looking for MarketplaceEntitlementClient and MarketplaceEntitlement.

Structs

Entitlement
An entitlement represents capacity in a product owned by the customer. For example, a customer might own some number of users or seats in an SaaS application or some amount of data capacity in a multi-tenant database.

EntitlementValue
The EntitlementValue represents the amount of capacity that the customer is entitled to for the product.

GetEntitlementsRequest
The GetEntitlementsRequest contains parameters for the GetEntitlements operation.

GetEntitlementsResult
The GetEntitlementsRequest contains results from the GetEntitlements operation.

MarketplaceEntitlementClient
A client for the AWS Marketplace Entitlement Service API.

Enums

GetEntitlementsError
Errors returned by GetEntitlements

Traits

MarketplaceEntitlement
Trait representing the capabilities of the AWS Marketplace Entitlement Service API. AWS Marketplace Entitlement Service clients implement this trait.